Prepartion for ESP32 support for generator control
