+++ include/Model/GeneratorForUno.hpp
@@ -1,25 +1,19 @@
 #ifndef GENRATORFORUNO_HPP
 #define GENRATORFORUNO_HPP
 
-#include "Model/GeneratorIf.hpp"
+#include "Model/GeneratorAd9833.hpp"
 
-class AD9833;
+#ifndef USE_ESP32
 
 /// @brief Class to generate waves with the AD9833 function generator
-class GeneratorForUno : public GeneratorIf
+/// for the Arduino Uno board.
+class GeneratorForUno : public GeneratorAd9833
 {
 
 public:
     GeneratorForUno(const uint8_t spiCsLine);
-    ~GeneratorForUno();
-    virtual void generateWave(const WaveType type, const long frequency);
-    virtual void disableWave();
-    virtual long getFrequency() const;
-    virtual WaveType getWaveType() const;
-    virtual bool isEnabled() const;
-
-private:
-    AD9833 *m_AD;
 };
+
+#endif
 
 #endif
