+++ src/Model/GeneratorForEsp32.cpp
@@ -1,0 +1,30 @@
+
+#ifdef ESP32
+
+GeneratorForEsp32(const uint8_t spiCsLine)
+{
+}
+~GeneratorForEsp32()
+{
+}
+void generateWave(const WaveType type, const long frequency);
+{
+}
+
+void disableWave()
+{
+}
+
+long getFrequency() const
+{
+}
+
+WaveType getWaveType() const
+{
+}
+
+bool isEnabled() const
+{
+}
+
+#endif
